prog run report: separate loading and rendering
the main method of the report generation was long.

- Refactor all code involved in loading data from the diags.nc
  files into a new module `prognostic_run/rundirs.json`.
- group plot generation and html generation into new "render" functions
- divide the html link generation into `find` and `render` phases
