+++ workflows/prognostic_run_diags/tests/_regtest_outputs/test_rundirs.test_render_links.out
@@ -1,0 +1,1 @@
+{'column_heating_moistening.mp4': "<a href='/Users/noah/workspace/VulcanClimateModeling/fv3net/workflows/prognostic_run_diags/45adebaeb631/run1/column_heating_moistening.mp4'>run1</a><a href='/Users/noah/workspace/VulcanClimateModeling/fv3net/workflows/prognostic_run_diags/45adebaeb631/run2/column_heating_moistening.mp4'>run2</a>"}
